Trying to simplify the gguf implementation and support more models
