+++ ext/candle/src/llm/mod.rs
@@ -6,10 +6,11 @@ pub mod llama;
 pub mod gemma;
 pub mod generation_config;
 pub mod text_generation;
-pub mod quantized;
+pub mod quantized_gguf;
 
 pub use generation_config::GenerationConfig;
 pub use text_generation::TextGeneration;
+pub use quantized_gguf::QuantizedGGUF;
 
 /// Trait for text generation models
 pub trait TextGenerator: Send + Sync {
