--- ext/candle/src/llm/quantized/mod.rs
@@ -1,7 +1,0 @@
-pub mod mistral;
-pub mod llama;
-pub mod gemma;
-
-pub use mistral::QuantizedMistral;
-pub use llama::QuantizedLlama;
-pub use gemma::QuantizedGemma;
